+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/Theme.TeumTeum"
         tools:targetApi="31">
         <activity
-            android:name=".ui.singup.SignUpActivity"
+            android:name=".ui.signup.SignUpActivity"
             android:exported="false" />
         <activity
             android:name=".ui.signin.LoginActivity"
