+++ app/src/main/java/com/example/teumteum/ui/signup/OnBoardingScheduleFragment.kt
@@ -1,4 +1,4 @@
-package com.example.teumteum.ui.singup
+package com.example.teumteum.ui.signup
 
 import BottomSheetScheduleFragment
 import android.os.Bundle
