+++ app/src/main/res/layout/fragment_complete.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@color/white"
-    tools:context=".ui.singup.CompleteFragment"
+    tools:context=".ui.signup.CompleteFragment"
     >
 
     <!-- 제목 -->
