+++ app/src/main/res/layout/fragment_on_boarding_nickname.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@color/white"
-    tools:context=".ui.singup.OnBoardingNicknameFragment">
+    tools:context=".ui.signup.OnBoardingNicknameFragment">
 
     <TextView
         android:id="@+id/title_tv"
