+++ app/src/main/res/layout/fragment_on_boarding_profile.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@color/white"
-    tools:context=".ui.singup.OnBoardingProfileFragment">
+    tools:context=".ui.signup.OnBoardingProfileFragment">
 
     <TextView
         android:id="@+id/title_tv"
