+++ app/src/main/res/layout/fragment_on_boarding_remind.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.singup.OnBoardingRemindFragment">
+    tools:context=".ui.signup.OnBoardingRemindFragment">
 
     <TextView
         android:id="@+id/title_tv"
