+++ app/src/main/res/layout/fragment_on_boarding_schedule.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.singup.OnBoardingScheduleFragment">
+    tools:context=".ui.signup.OnBoardingScheduleFragment">
 
     <TextView
         android:id="@+id/title_tv"
