+++ app/src/main/res/layout/fragment_on_boarding_sleep_pattern.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.singup.OnBoardingSleepPatternFragment">
+    tools:context=".ui.signup.OnBoardingSleepPatternFragment">
 
     <TextView
         android:id="@+id/title_tv"
