+++ app/src/main/res/layout/fragment_terms_detail.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.singup.TermsDetailFragment">
+    tools:context=".ui.signup.TermsDetailFragment">
 
 
     <TextView
